refactor(_gr.js): move logic into separate, imported functions
