+++ src/option_handlers/handleHelpOption.js
@@ -1,0 +1,6 @@
+// src/option_handlers/handleHelpOption.js
+
+export default function handleHelpOption(program) {
+  // Implicit exit code of 0
+  program.help();
+}
